fix: delete auto generated types
